flask_server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
boto3==1.34.88
botocore==1.34.88
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
lightgbm==4.3.0
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.23.2
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.0
pydantic_core==2.18.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
